fix bug memory

+++ Projet/src/operations.c
@@ -21,8 +21,8 @@
 #include <math.h>
 #include "operations.h"
 
-double calculerMoyenne(char *tableau[], int taille) {
-    double somme = 0.0;
+float calculerMoyenne(char *tableau[], int taille) {
+    float somme = 0.0;
 
     for (int i = 0; i < taille; i++) {
     
@@ -32,11 +32,11 @@ double calculerMoyenne(char *tableau[], int taille) {
     return taille > 0 ? somme / taille : 0.0;
 }
 
-double trouverMinimum(char *tableau[], int taille) {
-    double minimum = DBL_MAX;
+float trouverMinimum(char *tableau[], int taille) {
+    float minimum = DBL_MAX;
 
     for (int i = 0; i < taille; i++) {
-        double valeur = atof(tableau[i]);
+        float valeur = atof(tableau[i]);
         if (valeur < minimum) {
             minimum = valeur;
         }
@@ -45,11 +45,11 @@ double trouverMinimum(char *tableau[], int taille) {
     return minimum;
 }
 
-double trouverMaximum(char *tableau[], int taille) {
-    double maximum = -DBL_MAX;
+float trouverMaximum(char *tableau[], int taille) {
+    float maximum = -DBL_MAX;
 
     for (int i = 0; i < taille; i++) {
-        double valeur = atof(tableau[i]);
+        float valeur = atof(tableau[i]);
         if (valeur > maximum) {
             maximum = valeur;
         }
@@ -58,18 +58,18 @@ double trouverMaximum(char *tableau[], int taille) {
     return maximum;
 }
 
-double calculerEcartType(char *tableau[], int taille) {
-    double somme = 0.0;
-    double sommeCarres = 0.0;
+float calculerEcartType(char *tableau[], int taille) {
+    float somme = 0.0;
+    float sommeCarres = 0.0;
 
     for (int i = 0; i < taille; i++) {
-        double valeur = atof(tableau[i]);
+        float valeur = atof(tableau[i]);
         somme += valeur;
         sommeCarres += valeur * valeur;
     }
 
-    double moyenne = somme / taille;
-    double variance = (sommeCarres / taille) - (moyenne * moyenne);
+    float moyenne = somme / taille;
+    float variance = (sommeCarres / taille) - (moyenne * moyenne);
 
     return sqrt(variance);
 }
